requirements.txt
Django==2.2.4
django-mongodb-engine-py3==0.6.0.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
djangotoolbox==1.8.0
mongoengine==0.18.2
pymongo==3.9.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
